Add current gas limits to miner dashboard
